langgraph_test/requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio
aiohttp>=3.8.0
lxml>=4.9.0
feedparser>=6.0.0
arxiv>=2.0.0
pandas>=2.0.0
numpy>=1.24.0